% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Adds two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
Sum of \code{x} and \code{y}
}
\description{
Adds two numbers
}
\examples{
add(1,1)

}

