% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denplot.R
\name{denplot}
\alias{denplot}
\title{Plot posterior distributions from MCMC output}
\usage{
denplot(object, params = "all", ref_line = 0, ref_line_width = 1,
  quantiles = c(0.025, 0.975), rank = FALSE, xlim, xlab, main, labels,
  colors, dbar_height = 0.25, dbar_t_height = 0.35, dbar_t_width = 3,
  CI_t_color = "grey75")
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters to be
plotted. Partial names may be used to plot all parameters containing that set of characters.

Default \code{all} plots posteriors for all parameters. See VALUE below.

Valid entries are \code{jags_object}, \code{mcmc_list}, and \code{chains}. See DETAILS below.}

\item{ref_line}{Numerical vector indicating where vertical reference line should be created.

Default is \code{ref_line = 0}.

Argument \code{NULL} will plot no reference line.}

\item{ref_line_width}{Width of vertical reference lines.}

\item{quantiles}{Numerical vecor of length 2, indicating which quantiles to plot.

Default plots 95\% credible intervals.}

\item{rank}{If \code{TRUE} posteriors will ranked in decreasing order (based on
median posterior estimate) from top down.}

\item{xlim}{Numerical vector of length 2, indicating range of x-axis.}

\item{xlab}{Character string labeling x-axis.}

\item{main}{Character string indicating title of plot.}

\item{labels}{Character string (or vector of character strings if plotting > 1 parameter) labeling
parameter estimates along y-axis.

Specifying labels in the argument will use these to label density strips on y-axis.

Default option will use parameter names from \code{object}.

Option \code{NULL} will return plot with no labels on y-axis.}

\item{colors}{Vector of colors, indicating which color should should be used for each parameter. Default is 'black' for
all parameters.}

\item{dbar_height}{Height of density bar in plot.}

\item{dbar_t_height}{Height of ticks on density bar in plot.}

\item{dbar_t_width}{Width of ticks on density bar in plot.}

\item{CI_t_color}{Color of credible interval ticks.}
}
\description{
Plot the posterior distributions from MCMC output for specific parameters of interest using density
strips. Medians and credible intervals are indicated by ticks on density strips.
}
\section{Details}{

\code{object} argument can be an \code{mcmc.list} object, an \code{R2jags} model object (output from the \code{R2jags}
package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows
representing iterations in the chain).
}

\section{Notes}{

Plot code uses \code{denstrip} package, as highlighted in Jackson (2008) - generalized from code
for Zipkin et al. 2014, figure 3.
}

\section{References}{

Jackson, C. H. 2008. Displaying Uncertainty With Shading. The American Statistician 62:340-347.

Zipkin, E. F., T. S. Sillett, E. H. C. Grant, R. B. Chandler,
and J. A. Royle. 2014. Inferences about population dynamics
from count data using multistate models: a comparison to
capture-recapture approaches. Ecology and Evolution 4:417-426.
}
\examples{
#Load data
data(MCMC_data)

#Plot MCMC output
denplot(MCMC_data, labels=NULL)

#Just 'beta' parameters
denplot(MCMC_data, params= 'beta')

#Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
denplot(MCMC_data, params= c('beta[1]', 'gamma[4]', 'alpha[3]'))

#Rank parameters by posterior median
denplot(MCMC_data, params= 'beta', rank=TRUE)

}

