% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCMC_data}
\alias{MCMC_data}
\title{Simulated MCMC output data}
\format{\code{mcmc.list} object with 3 chains for each parameter, 1000 iterations for each chain}
\usage{
MCMC_data
}
\description{
Sample MCMC output containing 30 parameters - \code{alpha[1]}, ... , \code{alpha[10]},
\code{beta[1]}, ... , \code{beta[10]}, \code{gamma[1]}, ... , \code{gamma[10]}.
}
\keyword{datasets}

