% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\docType{package}
\name{post}
\alias{post}
\alias{post-package}
\title{The 'post' package}
\description{
The post package assists with plotting posterior
distributions, from MCMC chains.
}
\section{Details}{

At the moment, just the function 'poplot' is available
}

\section{Author(s)}{

Casey Youngflesh <casey.youngflesh@stonybrook.edu>
}

