% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posummary.R
\name{posummary}
\alias{posummary}
\title{Summary function for MCMC output}
\usage{
posummary(object, par = "all", excl = NULL, Rhat = TRUE)
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{par}{Character string (or vector of character strings) denoting parameters to be
returned in summary output. Partial names may be used to return all parameters containing
that set of characters.

Default \code{'all'} returns all parameters in summary output.}

\item{excl}{Character string (or vector of character strings) denoting parameters to exclude.
Partical names may be used to exclude all parameters contaiing that set of characters. Used in
conjunction with \code{par} argument to select parameters of interest.}

\item{Rhat}{If \code{TRUE}, summary information contains Gelman-Rubin convergence statistic (Rhat)
and if \code{FALSE}, Rhat output is masked.}
}
\value{
Function returns summary information (including parameter posterior mean, 2.5\% quantile, median, 97.5\%
 quantile, and Gelman-Rubin convergence statistic (Rhat)) for specified parameters.
}
\description{
Extract summary information from MCMC output for specific parameters of interest.
}
\section{Details}{

\code{object} argument can be a \code{stanfit} object (\code{rstan} package), an \code{mcmc.list} object
(\code{coda} package), an \code{R2jags} model object (\code{R2jags} package), or a matrix containing MCMC
chains (each column representing MCMC output for a single parameter, rows representing iterations in the chain).
}

\section{Notes}{


For \code{mcmc.list} objects, Gelman-Rubin convergence statistic  (Rhat) is calculated using the
\code{gelman.diag} function in the \code{coda} package.
}
\examples{
#Load data
data(MCMC_data)

#Summary information for MCMC output
posummary(MCMC_data)

#Just 'beta' parameters
posummary(MCMC_data, par= 'beta')

#Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
posummary(MCMC_data, par= c('beta[1]', 'gamma[4]', 'alpha[3]'))

}

