% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_data.R
\docType{data}
\name{MCMC_data}
\alias{MCMC_data}
\title{Simulated MCMC output data}
\format{
\code{mcmc.list} object with 3 chains for each parameter, 6000 iterations for each chain.
}
\usage{
MCMC_data
}
\description{
Sample MCMC output containing 12 parameters - \code{alpha[1]}, ... , \code{alpha[6]}, \code{beta[1]}, ... , \code{beta[6]}.
}
\keyword{datasets}
