% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pochains.R
\name{pochains}
\alias{pochains}
\title{Extract posterior chains from MCMC output}
\usage{
pochains(object, params = "all")
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters of interest.
Partial names may be used to return all parameters containing that set of characters.

Default \code{'all'} returns chains for all parameters.}
}
\description{
Extract posterior chains from MCMC output for specific parameters of interest.
}
\section{Details}{

Function returns matrix with one chain per column for specified parameters. Multiple input chains for each
parameter are combined to one posterior chain.

\code{object} argument can be a \code{stanfit} object (\code{rstan} package), an \code{mcmc.list} object
(\code{coda} package), an \code{R2jags} model object (\code{R2jags} package), or a matrix containing MCMC
chains (each column representing MCMC output for a single parameter, rows representing iterations in the chain).
}
\examples{
#Load data
data(MCMC_data)

#Extract MCMC chains
ex <- pochains(MCMC_data)
apply(ex, 2, mean)

#Extract MCMC chains for just 'beta' parameters
ex2 <- pochains(MCMC_data, params='beta')
apply(ex2, 2, mean)

}

