% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pochains.R
\name{pochains}
\alias{pochains}
\title{Chain extraction for jags object}
\usage{
pochains(object, params = "all")
}
\arguments{
\item{object}{JAGS model object - output from JAGS model}

\item{params}{Character string (or vector of character strings) denoting parameters to be
returned in summary output. Partial names may be used to return all parameters containing
that set of characters. Default \code{'all'} returns all parameters in summary output.}
}
\value{
\code{pochains(params='all')} returns MCMC chains for all parameters contained within
JAGS model object.

\code{pochains(params=c('beta[1]', 'beta[2]'))} returns MCMC chains for just parameters
\code{'beta[1]'} and \code{'beta[2]'}.

\code{pochains(params=c('beta'))} returns MCMC chains for all parameters containing \code{'beta'}
 in their name.
}
\description{
Extracts posterior MCMC chains from JAGS object output by JAGS model fit in R
}
\section{Details}{

Function currently only supports JAGS model objects output by \code{R2jags} package.
}

