% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pochains.R
\name{pochains}
\alias{pochains}
\title{Chain extraction for jags object}
\usage{
pochains(jags_object, params = "all")
}
\arguments{
\item{jags_object}{Output from jags model}

\item{params}{Parameters of interest from jags object. Partial names
may be used to return all object containing that name.}
}
\value{
Chains for particular parameters of interest
}
\description{
Simplifies extracting particular chains from jags object
}
\section{Details}{

Output from R2jags only right now.
}

