% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplot.R
\name{poplot}
\alias{poplot}
\title{Plots posterior distributions}
\usage{
poplot(data, LML = -1, LMH = 1)
}
\arguments{
\item{data}{Matrix of MCMC chains}

\item{LML}{Lower bounds x-axis}

\item{LMH}{upper bounds x-axis}
}
\description{
Plots posterior distributions
}
\section{Details}{

Chains should be in columns of input matrix
}

