% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplot.R
\name{poplot}
\alias{poplot}
\title{Plot posterior distributions from MCMC output}
\usage{
poplot(object,
        params = "all",
        input = "jags_object",
        quantiles = c(0.025, 0.975),
        centrality = "mean",
        g_lines = 0,
        xlim,
        xlab = "Parameter probability values",
        ylab,
        main,
        dbar_height = 0.2,
        tick_height = 0.5,
        tick_width = 3)
}
\arguments{
\item{object}{Object containing MCMC output. See \code{input} argument and DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters to be
plotted. Partial names may be used to plot all parameters containing that set of characters.

Default \code{all} plots posteriors for all parameters. See VALUE below.}

\item{input}{Indicates the nature of the \code{object} argument.

Valid entries are \code{jags_object}, \code{mcmc_list}, and \code{chains}. See DETAILS below.}

\item{quantiles}{Numerical vecor of length 2, indicating which quantiles to plot.

Default plots 95\% credible intervals.}

\item{centrality}{Indicates which measure of centrality to plot.

Valid options are \code{mean}
and \code{median}.}

\item{g_lines}{Numerical vector indicating where vertical reference lines should be created.

Default is \code{g_lines = 0}.

Argument \code{NULL} will plot no guidelines.}

\item{xlim}{Numerical vector of length 2, indicating range of x-axis.}

\item{xlab}{Character string labeling x-axis.}

\item{ylab}{Character string (or vector of character strings if plotting > 1 parameter) labeling
y-axis.

Specifying labels in the argument will use these to label axis.

Default option will use parameter names from \code{object}.

Option \code{NULL} will return plot with no labels on y-axis.}

\item{main}{Character string indicating title of plot.}

\item{dbar_height}{Height of density bar in plot.}

\item{tick_height}{Height of ticks in plot.}

\item{tick_width}{Width of ticks in plot.}
}
\value{
\code{posummary(params = 'all')} returns posterior plots for all parameters contained within
JAGS model object.

\code{posummary(params = c('beta[1]', 'beta[2]'))} returns posterior plots for just parameters
\code{beta[1]} and \code{beta[2]}.

\code{posummary(params = 'beta')} returns posterior plots for all parameters containing \code{beta}
 in their name.
}
\description{
Plot the posterior distributions from MCMC output for specific parameters of interest. All posterior
parameter estimates are plotted on the same plot using density strips, similar to a caterpillar plot.
}
\section{Details}{

For \code{posummary(object, input = 'jags_object')}, input must be JAGS model object from \code{R2jags} package.

For \code{posummary(object, input = 'mcmc_list')}, input must be of type \code{mcmc.list}.

For \code{posummary(object, input = 'chains')}, each column of \code{object} should contain a posterior
chain for a single parameter. Each row represents one iteration in the chain.
}

\section{Notes}{

Plot code uses \code{denstrip} package, as highlighted in Jackson (2008) - generalized from code
for Zipkin et al. 2014, figure 3.
}

\section{References}{

Jackson, C. H. 2008. Displaying Uncertainty With Shading. The American Statistician 62:340-347.

Zipkin, E. F., T. S. Sillett, E. H. C. Grant, R. B. Chandler,
and J. A. Royle. 2014. Inferences about population dynamics
from count data using multistate models: a comparison to
capture-recapture approaches. Ecology and Evolution 4:417-426.
}
\examples{
x1 <- rnorm(1000, mean=0.5)
x2 <- rnorm(1000, mean=0)
data <- cbind(x1, x2)
poplot(data, input = 'chains')

}

