% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplot.R
\name{poplot}
\alias{poplot}
\title{Plot posterior distributions from MCMC output - VERSION 2}
\usage{
poplot(object, params = "all", ref_line = 0, rank = FALSE, xlim, ylim,
  xlab, main, labels, labels_sz = 1.2, med_sz = 1.5, thick_sz = 5,
  thin_sz = 2, ax_sz = 3, x_axis_text_sz = 1.3, x_tick_text_sz = 1.2,
  main_text_sz = 1, tick_pos, mar = c(5.1, 4.1, 4.1, 2.1))
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters to be
plotted. Partial names may be used to plot all parameters containing that set of characters.

Default \code{all} plots posteriors for all parameters. See VALUE below.

Valid entries are \code{jags_object}, \code{mcmc_list}, and \code{chains}. See DETAILS below.}

\item{ref_line}{Numerical vector indicating where vertical reference line should be created.

Default is \code{ref_line = 0}.

Argument \code{NULL} will plot no reference line.}

\item{rank}{If \code{TRUE} posteriors will ranked in decreasing order (based on
specified measure of centrality) from top down.}

\item{xlim}{Numerical vector of length 2, indicating range of x-axis.}

\item{ylim}{Numerical vector of length 2, indicating range of y-axis}

\item{xlab}{Character string labeling x-axis.

Option \code{NULL} will return plot with no labels on y-axis.}

\item{main}{Character string indicating title of plot.}

\item{labels}{Character string (or vector of character strings if plotting > 1 parameter) labeling
parameter estimates along y-axis.

Specifying labels in the argument will use these to label parameter estimates on y-axis.

Default option will use parameter names from \code{object}.

Option \code{NULL} will return plot with no labels on y-axis.}

\item{labels_sz}{Number specifying size of text for parameter labels on y-axis.}

\item{med_sz}{Number specifying size of points represents posterior medians.}

\item{thick_sz}{Number specifying thickness of 50 percent CI line (thicker line).}

\item{thin_sz}{Number specifying thickness of 95 percent CI line (thinner line).}

\item{ax_sz}{Number specifying thickness of x-axis and ticks.}

\item{x_axis_text_sz}{Number specifying size of text for x-axis label.}

\item{x_tick_text_sz}{Number specifying size of text for tick labels on x-axis.}

\item{main_text_sz}{Number specifying size of text for main title.}

\item{tick_pos}{Numeric vector specifying where ticks on x-axis should be placed.}

\item{mar}{Numerical vector of length 4 specifying plot margins - (BOTTOM, LEFT, TOP, RIGHT).
Changes to the margin should be made within the function rather than using the \code{par} call.

Default is c(5.1, 4.1, 4.1, 2.1) - the R plot default.}
}
\description{
Plot the posterior distributions from MCMC output for specific parameters of interest. All posterior
parameter estimates are plotted as a caterpillar plot.
}
\section{Details}{

Points represent posterior medians. Parameters
which 50 percent credible intervals overlap 0 are indicated by 'open' circles. Parameters which 50 percent credible
intervals do not overlap 0 and 95 percent credible intervals do overlap 0 are indicated by 'closed' grey circles.
Parameters which 95 percent credible intervals do not overlap 0 are indicated by 'closed' black circles. Thick
lines represent 50 percent credible intervals while thin lines represent 95 percent credible intervals.

\code{object} argument can be an \code{mcmc.list} object, an \code{R2jags} model object (output from the \code{R2jags}
package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows
representing iterations in the chain).
}

\section{Notes}{


When specifying \code{rank = TRUE} and specifying labels for \code{ylab}, labels will be applied to parameters before
they are ranked.

Thanks to Cinner et al. 2016, whose Fig. 1 inspired this plot.
}

\section{References}{


Cinner, J. E., C. Huchery, M. A. MacNeil, N. A. J. Graham, T. R. McClanahan, J. Maina, E. Maire, J.
N. Kittinger, C. C. Hicks, C. Mora, E. H. Allison, S. D’Agata, A. Hoey, D. A. Feary, L. Crowder, I.
D. Williams, M. Kulbicki, L. Vigliola, L. Wantiez, G. Edgar, R. D. Stuart-Smith, S. A. Sandin, A.
L. Green, M. J. Hardt, M. Beger, A. Friedlander, S. J. Campbell, K. E. Holmes, S. K. Wilson, E.
Brokovich, A. J. Brooks, J. J. Cruz-Motta, D. J. Booth, P. Chabanet, C. Gough, M. Tupper, S. C. A.
Ferse, U. R. Sumaila, and D. Mouillot. 2016. Bright spots among the world’s coral reefs. Nature
535:416-419.
}
\examples{
#Load data
data(MCMC_data)

#Plot MCMC output
poplot(MCMC_data, ylab=NULL)

#Just 'beta' parameters
poplot(MCMC_data, params= 'beta')

#Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
poplot(MCMC_data, params= c('beta[1]', 'gamma[4]', 'alpha[3]'))

#Rank parameters by posterior mean
poplot(MCMC_data, params= 'beta', rank=TRUE)

}

