% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poplot.R
\name{poplot}
\alias{poplot}
\title{Plot posterior distributions from MCMC output}
\usage{
poplot(object, params = "all", g_lines = 0, quantiles = c(0.025, 0.975),
  rank = FALSE, centrality = "mean", xlim,
  xlab = "Parameter probability values", ylab, main, dbar_height = 0.25,
  dbar_t_height = 0.35, dbar_t_width = 3)
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters to be
plotted. Partial names may be used to plot all parameters containing that set of characters.

Default \code{all} plots posteriors for all parameters. See VALUE below.

Valid entries are \code{jags_object}, \code{mcmc_list}, and \code{chains}. See DETAILS below.}

\item{g_lines}{Numerical vector indicating where vertical reference lines should be created.

Default is \code{g_lines = 0}.

Argument \code{NULL} will plot no guidelines.}

\item{quantiles}{Numerical vecor of length 2, indicating which quantiles to plot.

Default plots 95\% credible intervals.}

\item{rank}{If \code{TRUE} posteriors will ranked in decreasing order (based on
specified measure of centrality) from top down.}

\item{centrality}{Indicates which measure of centrality to plot.

Valid options are \code{mean} and \code{median}.}

\item{xlim}{Numerical vector of length 2, indicating range of x-axis.}

\item{xlab}{Character string labeling x-axis.}

\item{ylab}{Character string (or vector of character strings if plotting > 1 parameter) labeling
y-axis.

Specifying labels in the argument will use these to label axis.

Default option will use parameter names from \code{object}.

Option \code{NULL} will return plot with no labels on y-axis.}

\item{main}{Character string indicating title of plot.}

\item{dbar_height}{Height of density bar in plot.}

\item{dbar_t_height}{Height of ticks on density bar in plot.}

\item{dbar_t_width}{Width of ticks on density bar in plot.}
}
\value{
Function returns density strip plot, similar to caterpillar plot, for all specified parameters. Plotted output
can be sorted by mean estimate.
}
\description{
Plot the posterior distributions from MCMC output for specific parameters of interest. All posterior
parameter estimates are plotted on the same plot using density strips, similar to a caterpillar plot.
}
\section{Details}{

\code{object} argument can be an \code{mcmc.list} object, an \code{R2jags} model object (output from the \code{R2jags}
package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows
representing iterations in the chain).
}

\section{Notes}{

Plot code uses \code{denstrip} package, as highlighted in Jackson (2008) - generalized from code
for Zipkin et al. 2014, figure 3.
}

\section{References}{

Jackson, C. H. 2008. Displaying Uncertainty With Shading. The American Statistician 62:340-347.

Zipkin, E. F., T. S. Sillett, E. H. C. Grant, R. B. Chandler,
and J. A. Royle. 2014. Inferences about population dynamics
from count data using multistate models: a comparison to
capture-recapture approaches. Ecology and Evolution 4:417-426.
}
\examples{
#Load data
data(MCMC_data)

#Plot MCMC output
poplot(MCMC_data, ylab=NULL)

#Just 'beta' parameters
poplot(MCMC_data, params= 'beta')

#Just 'beta[1]', 'gamma[4]', and 'alpha[3]'
poplot(MCMC_data, params= c('beta[1]', 'gamma[4]', 'alpha[3]'))

#Rank parameters by posterior mean
poplot(MCMC_data, params= 'beta', rank=TRUE)

}

