% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\docType{package}
\name{post}
\alias{post}
\alias{post-package}
\title{The 'post' package}
\description{
The post package assists with summarizing posterior MCMC chains
(including parameter estimate extraction, chain extraction, and
visualizing posterior distributions), obtained from Bayesian model
ouput.
}
\section{Details}{

The following functions are currently available:

     -posummary (summarizes jags model output)

     -pochains (extracts chains from jags model output)

     -poplot (plots posterior chains from jags model output)
}

\section{Workflow}{

Workflow of package as follows:

     -posummary(jags_object, params)

     -chains <- pochains(jags_object, params)

     -poplot(chains)
}

\section{Author(s)}{

Casey Youngflesh <casey.youngflesh@stonybrook.edu>
}

