% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posummary.R
\name{posummary}
\alias{posummary}
\title{Summary function for jags object}
\usage{
posummary(jags_object, params = "all")
}
\arguments{
\item{jags_object}{Output from jags model}

\item{params}{Parameters of interest from jags object. Partial names
may be used to return all object containing that name.}
}
\value{
Summary data (mean, 95\% quantile, R_hat) for particular parameters of interest
}
\description{
Simplifies extracting info for particular parameters from jags object
}
\section{Details}{

Output from R2jags only right now.
}

