% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posummary.R
\name{posummary}
\alias{posummary}
\title{Summary function for MCMC output}
\usage{
posummary(object, params = "all", Rhat = TRUE)
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters to be
returned in summary output. Partial names may be used to return all parameters containing
that set of characters.

Default \code{all} returns all parameters in summary output.}

\item{Rhat}{If \code{TRUE}, summary information contains Gelman-Rubin convergence statistic (Rhat)
and if \code{FALSE}, Rhat output is masked.}
}
\value{
\code{posummary(params='all')} returns summary data for all parameters contained within
JAGS model object.

\code{posummary(params=c('beta[1]', 'beta[2]'))} returns summary data for just parameters
\code{beta[1]} and \code{beta[2]}.

\code{posummary(params=c('beta'))} returns summary data for all parameters containing \code{beta}
 in their name.
}
\description{
Extract summary information from MCMC output for specific parameters of interest.
}
\section{Details}{

\code{object} argument can be an \code{mcmc.list} object, an \code{R2jags} model object (output from the \code{R2jags}
package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows
representing iterations in the chain).
}

\section{Notes}{

 Default summary information includes (parameter posterior mean, 2.5\% quantile, median, 97.5\%
 quantile, and Gelman-Rubin convergence statistic (Rhat)).

For \code{mcmc.list} objects, Gelman-Rubin convergence statistic  (Rhat) is calculated using the
\code{gelman.diag} function in the \code{coda} package.
}

