% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potrace.R
\name{potrace}
\alias{potrace}
\title{Plot MCMC chains to check for convergence}
\usage{
potrace(object, params = "all")
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters of interest.
Partial names may be used to return all parameters containing that set of characters.

Default \code{all} returns chains for all parameters.}
}
\value{
\code{potrace(params='all')} returns chains for all parameters.

\code{potrace(params=c('beta[1]', 'beta[2]'))} returns chains for just parameters \code{beta[1]} and \code{beta[2]}.

\code{potrace(params=c('beta'))} returns chains for all parameters containing \code{beta} in their name.
}
\description{
Plot MCMC chains for specific parameters of interest.
}
\section{Details}{

\code{object} argument can be an \code{mcmc.list} object, an \code{R2jags} model object (output from the \code{R2jags}
package), or a matrix containing MCMC chains (each column representing MCMC output for a single parameter, rows
representing iterations in the chain).
}

