% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potrace.R
\name{potrace}
\alias{potrace}
\title{Plot MCMC chains to check for convergence}
\usage{
potrace(object, params = "all", pdf = FALSE, wd = getwd())
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters of interest.
Partial names may be used to return all parameters containing that set of characters.

Default \code{all} returns chains for all parameters.}

\item{pdf}{Logical - if \code{pdf = TRUE} plots will be exported to a pdf.}

\item{wd}{Working directory for pdf output. Default is current directory.}
}
\value{
Function returns trace plots for specified parameters. Output can be printed to pdf if specified.
}
\description{
Plot MCMC chains for specific parameters of interest. Option to print trace plots
to pdf output, which greater increases review efficiency with large numbers of
parameters.
}
\section{Details}{

\code{object} argument can be an \code{mcmc.list} object or an \code{R2jags} model object (output from the \code{R2jags}
package).
}
\examples{
#Load data
data(MCMC_data)

#Traceplot for all 'beta' parameters
potrace(MCMC_data, params='beta')

#Print traceplot output to pdf
potrace(MCMC_data, pdf= TRUE)

}

