% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potrace.R
\name{potrace}
\alias{potrace}
\title{Trace and density plots from MCMC output}
\usage{
potrace(object, params = "all", pdf = FALSE, filename, wd = getwd(),
  type = "both")
}
\arguments{
\item{object}{Object containing MCMC output. See DETAILS below.}

\item{params}{Character string (or vector of character strings) denoting parameters of interest.
Partial names may be used to return all parameters containing that set of characters.

Default \code{'all'} returns chains for all parameters.}

\item{pdf}{Logical - if \code{pdf = TRUE} plots will be exported to a pdf.}

\item{filename}{Name of pdf file to be printed.}

\item{wd}{Working directory for pdf output. Default is current directory.}

\item{type}{Type of plot to be output. \code{'both'} outputs both trace and density plots, \code{'trace'}
outputs only trace plots, and \code{'density'} outputs only density plots.}
}
\description{
Trace plots and density plots of MCMC chains for specific parameters of interest. Option to
print plots to pdf.
}
\section{Details}{

Plots created similar to that of \code{traceplot} from \code{coda} package.

\code{object} argument can be a \code{stanfit} object (\code{rstan} package), an \code{mcmc.list} object
(\code{coda} package), or an \code{R2jags} model object (\code{R2jags} package).
}
\examples{
#Load data
data(MCMC_data)

#Traceplot for all 'beta' parameters
potrace(MCMC_data, params='beta')

#Print traceplot output to pdf
potrace(MCMC_data, pdf= TRUE, filename = 'PDF_file.pdf')

}

